+++ Fun.Build/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add indicator for default pipeline
+
 ## [1.0.6] - 2024-01-17
 
 - Fix typo